requirements.txt
requests==2.32.4
python-dotenv==1.1.1
rich==14.1.0
typing-extensions==4.14.1
fastapi==0.116.1
pydantic==2.11.7 